requirements.txt
streamlit==1.25.0
streamlit-drawable-canvas==0.9.3
Pillow==9.5.0
fpdf2==2.8.4
google-generativeai==0.8.5
elevenlabs==0.2.9
numpy==1.26.4
pandas==2.3.2
typing_extensions==4.15.0
